Suppress backlog clutter
* Overwrite Streamlink's output after every call that doesn't return an exit code of 0.
This should preserve Streamlink's output when a live stream has been detected.
For every other call where a live stream is not active, we safely overwrite the repeating output and keep only the last two lines.
To achieve this we use special non-printable characters to tell the terminal to go back two columns up, clear the line to the beginning, write the date, and clear the rest of the line.
The special characters might be different from one terminal emulator to another, so we ask tput for that information if available.
* Choose a random sleep value to slightly reduce risk of getting banned by twitch.